// src/app/modules/map/models/country.ts
export class Country {

    /** Country Name */
    Country: string;

    /** Country Code */
    CountryCode: string;

    /** slug */
    Slug: string;

    /** New Confirmed cases */
    NewConfirmed: number;

    /** Total of Confirmed cases */
    TotalConfirmed: number;

    /** New Deaths */
    NewDeaths: number;

    /** Total Deaths */
    TotalDeaths: number;

    /** New Recovered cases */
    NewRecovered: number;

    /** Total Recovered cases */
    TotalRecovered: number;

    Lat: number;

    Lon: number;

    /** Date */
    Date: Date;

}